split out simulation api to internal api, update tests to use correct fees flow
