+++ extension/src/popup/ducks/token-payment.ts
@@ -1,7 +1,7 @@
 import { Address, SorobanRpc, XdrLargeInt } from "stellar-sdk";
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import { ActionStatus, ErrorMessage } from "@shared/api/types";
-import { INDEXER_URL } from "@shared/constants/mercury";
+import { simulateTokenTransfer } from "@shared/api/internal";
 import { NetworkDetails } from "@shared/constants/stellar";
 import { SorobanRpcNotSupportedError } from "@shared/constants/errors";
 import { transfer } from "@shared/helpers/soroban/token";
@@ -78,35 +78,21 @@ export const simulateTokenPayment = createAsyncThunk<
           preparedTransaction,
         };
       }
-      const options = {
-        method: "POST",
-        headers: {
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          address,
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          pub_key: publicKey,
-          memo,
-          params,
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          network_url: networkDetails.sorobanRpcUrl,
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          network_passphrase: networkDetails.networkPassphrase,
-        }),
-      };
-      const res = await fetch(
-        `${INDEXER_URL}/simulate-token-transfer`,
-        options,
-      );
-      const response = await res.json();
+      const { ok, response } = await simulateTokenTransfer({
+        address,
+        publicKey,
+        memo,
+        params,
+        networkDetails,
+        transactionFee,
+      });
 
-      if (!res.ok) {
+      if (!ok) {
         return thunkApi.rejectWithValue({
           errorMessage: response.message,
         });
       }
+
       return {
         preparedTransaction: response.preparedTransaction,
         simulationTransaction: response.simulationResponse,
